Refactor item state management: remove legacy isDone property, streamline state handling in GroceryItem class
